se finaliza proyecto

+++ admin/secciones/entradas/editar.php
@@ -16,55 +16,58 @@ if (isset($_GET['txtID'])) {
 }
 
 if($_POST){
-
-    $txtID = (isset($_POST['txtID'])) ? $_POST['txtID'] : "";
-    $fecha = (isset($_POST['fecha'])) ? $_POST['fecha'] : "";
-    $titulo = (isset($_POST['titulo'])) ? $_POST['titulo'] : "";
-    $descripcion = (isset($_POST['descripcion'])) ? $_POST['descripcion'] : "";
-    
-    $sentencia = $conexion->prepare("UPDATE `tbl_entradas`
-    SET fecha = :fecha, titulo = :titulo, descripcion = :descripcion WHERE ID = :ID");
-
-    $sentencia->bindParam(':fecha', $fecha);
-    $sentencia->bindParam(':titulo', $titulo);
-    $sentencia->bindParam(':descripcion', $descripcion);
-    $sentencia->bindParam(':ID', $txtID);
-    $sentencia->execute();
-    $mensaje = "Datos agregados correctamente";
-    header("Location: index.php?mensaje=$mensaje");
-
-    if ($_FILES['imagen']['tmp_name'] != "") {
-
-        $imagen = (isset($_FILES['imagen']['name'])) ? $_FILES['imagen']['name'] : "";
-        $fecha_imagen = new DateTime();
-        $nombre_archivo_imagen = ($imagen != "") ? $fecha_imagen->getTimestamp() . "_" . $imagen : "";
-
-        $tmp_imagen = $_FILES["imagen"]["tmp_name"];
-
-        move_uploaded_file($tmp_imagen, "../../../assets/img/about/" . $nombre_archivo_imagen);
+    // Validar que los campos requeridos no estén vacíos
+    if (empty($_POST['fecha']) || empty($_POST['titulo']) || empty($_POST['descripcion'])) {
+        $mensaje = "Por favor, complete todos los campos requeridos";
+    } else {
+        $txtID = (isset($_POST['txtID'])) ? $_POST['txtID'] : "";
+        $fecha = (isset($_POST['fecha'])) ? $_POST['fecha'] : "";
+        $titulo = (isset($_POST['titulo'])) ? $_POST['titulo'] : "";
+        $descripcion = (isset($_POST['descripcion'])) ? $_POST['descripcion'] : "";
         
-        //borrado del archivo anterior
-        $sentencia = $conexion->prepare("SELECT imagen FROM `tbl_entradas` WHERE ID = :ID");
+        $sentencia = $conexion->prepare("UPDATE `tbl_entradas`
+        SET fecha = :fecha, titulo = :titulo, descripcion = :descripcion WHERE ID = :ID");
+
+        $sentencia->bindParam(':fecha', $fecha);
+        $sentencia->bindParam(':titulo', $titulo);
+        $sentencia->bindParam(':descripcion', $descripcion);
         $sentencia->bindParam(':ID', $txtID);
         $sentencia->execute();
-        $registro = $sentencia->fetch(PDO::FETCH_LAZY);
-
-        if (isset($registro["imagen"])) {
-            if(file_exists("../../../assets/img/about/" . $registro["imagen"])) {
-                unlink("../../../assets/img/about/" . $registro["imagen"]);
-            }
-        }
-
-        $sentencia = $conexion->prepare("UPDATE `tbl_entradas` SET `imagen` = :imagen WHERE `ID` = :ID;");
-        $sentencia->bindParam(':imagen', $nombre_archivo_imagen);
-        $sentencia->bindParam(':ID', $txtID);
-        $sentencia->execute();
-        $imagen=$nombre_archivo_imagen;
         $mensaje = "Datos agregados correctamente";
         header("Location: index.php?mensaje=$mensaje");
+
+        if ($_FILES['imagen']['tmp_name'] != "") {
+
+            $imagen = (isset($_FILES['imagen']['name'])) ? $_FILES['imagen']['name'] : "";
+            $fecha_imagen = new DateTime();
+            $nombre_archivo_imagen = ($imagen != "") ? $fecha_imagen->getTimestamp() . "_" . $imagen : "";
+
+            $tmp_imagen = $_FILES["imagen"]["tmp_name"];
+
+            move_uploaded_file($tmp_imagen, "../../../assets/img/about/" . $nombre_archivo_imagen);
+            
+            //borrado del archivo anterior
+            $sentencia = $conexion->prepare("SELECT imagen FROM `tbl_entradas` WHERE ID = :ID");
+            $sentencia->bindParam(':ID', $txtID);
+            $sentencia->execute();
+            $registro = $sentencia->fetch(PDO::FETCH_LAZY);
+
+            if (isset($registro["imagen"])) {
+                if(file_exists("../../../assets/img/about/" . $registro["imagen"])) {
+                    unlink("../../../assets/img/about/" . $registro["imagen"]);
+                }
+            }
+
+            $sentencia = $conexion->prepare("UPDATE `tbl_entradas` SET `imagen` = :imagen WHERE `ID` = :ID;");
+            $sentencia->bindParam(':imagen', $nombre_archivo_imagen);
+            $sentencia->bindParam(':ID', $txtID);
+            $sentencia->execute();
+            $imagen=$nombre_archivo_imagen;
+            $mensaje = "Datos agregados correctamente";
+            header("Location: index.php?mensaje=$mensaje");
+        }
     }
 }
-
 
 include("../../templates/headers.php");
 ?>
@@ -77,25 +80,24 @@ include("../../templates/headers.php");
 
         <form action="" method="post" enctype="multipart/form-data">
 
-        <div class="mb-3">
-          <label for="" class="form-label">ID:</label>
-          <input type="text"
-            class="form-control" readonly value="<?php echo $txtID;?>" name="txtID" id="txtID" aria-describedby="helpId" placeholder="">
-          </div>
+            <div class="mb-3">
+                <label for="" class="form-label">ID:</label>
+                <input type="text" class="form-control" readonly value="<?php echo $txtID;?>" name="txtID" id="txtID" aria-describedby="helpId" placeholder="">
+            </div>
 
             <div class="mb-3">
                 <label for="fecha" class="form-label">Fecha:</label>
-                <input type="date" class="form-control" value="<?php echo $fecha;?>" name="fecha" id="fecha" aria-describedby="helpId" placeholder="Fecha">
+                <input type="date" class="form-control" value="<?php echo $fecha;?>" name="fecha" id="fecha" aria-describedby="helpId" placeholder="Fecha" required>
             </div>
 
             <div class="mb-3">
                 <label for="titulo" class="form-label">Titulo:</label>
-                <input type="text" class="form-control" value="<?php echo $titulo;?>" name="titulo" id="titulo" aria-describedby="helpId" placeholder="Titulo">
+                <input type="text" class="form-control" value="<?php echo $titulo;?>" name="titulo" id="titulo" aria-describedby="helpId" placeholder="Titulo" required>
             </div>
 
             <div class="mb-3">
                 <label for="descripcion" class="form-label">Descripcion: </label>
-                <input type="text" class="form-control" value="<?php echo $descripcion;?>" name="descripcion" id="descripcion" aria-describedby="helpId" placeholder="Descripcion">
+                <input type="text" class="form-control" value="<?php echo $descripcion;?>" name="descripcion" id="descripcion" aria-describedby="helpId" placeholder="Descripcion" required>
             </div>
 
             <div class="mb-3">
@@ -106,13 +108,15 @@ include("../../templates/headers.php");
 
             <button type="submit" class="btn btn-success">Actualizar</button>
             <a name="" id="" class="btn btn-primary" href="index.php" role="button">Cancelar</a>
+
+            <?php if (isset($mensaje)) : ?>
+                <div class="alert alert-danger mt-3" role="alert">
+                    <?php echo $mensaje; ?>
+                </div>
+            <?php endif; ?>
         </form>
-
-
-
     </div>
     <div class="card-footer text-muted">
-
     </div>
 </div>
