+++ admin/secciones/equipo/editar.php
@@ -18,80 +18,78 @@ if (isset($_GET['txtID'])) {
 
 if ($_POST) {
     $txtID = (isset($_POST['txtID'])) ? $_POST['txtID'] : "";
-    $imagen = (isset($_FILES['imagen']["name"])) ? $_FILES['imagen']["name"] : "";
+    $imagen = (isset($_FILES['imagen']['name'])) ? $_FILES['imagen']['name'] : "";
     $nombrecompleto = (isset($_POST['nombrecompleto'])) ? $_POST['nombrecompleto'] : "";
     $puesto = (isset($_POST['puesto'])) ? $_POST['puesto'] : "";
     $twitter = (isset($_POST['twitter'])) ? $_POST['twitter'] : "";
     $facebook = (isset($_POST['facebook'])) ? $_POST['facebook'] : "";
     $linkedin = (isset($_POST['linkedin'])) ? $_POST['linkedin'] : "";
 
-    $sentencia = $conexion->prepare("UPDATE `tbl_equipo` SET
-    `nombrecompleto` = :nombrecompleto,`puesto` = :puesto,`twitter` = :twitter,`facebook` = :facebook,`linkedin` = :linkedin
-    WHERE `ID` = :ID;");
+    // Validar campos requeridos
+    if (empty($nombrecompleto) || empty($puesto)) {
+        $mensaje = "Por favor, complete los campos requeridos";
+    } else {
+        $sentencia = $conexion->prepare("UPDATE `tbl_equipo` SET
+        `nombrecompleto` = :nombrecompleto, `puesto` = :puesto, `twitter` = :twitter, `facebook` = :facebook, `linkedin` = :linkedin
+        WHERE `ID` = :ID;");
        
-    $sentencia->bindParam(':nombrecompleto', $nombrecompleto);
-    $sentencia->bindParam(':puesto', $puesto);
-    $sentencia->bindParam(':twitter', $twitter);
-    $sentencia->bindParam(':facebook', $facebook);
-    $sentencia->bindParam(':linkedin', $linkedin);
-    $sentencia->bindParam(':ID', $txtID);
-    $sentencia->execute();
-    $mensaje = "Datos actualizados correctamente";
-    header("Location: index.php?mensaje=$mensaje");
-
-
-    if ($_FILES['imagen']['tmp_name'] != "") {
-
-        $imagen = (isset($_FILES['imagen']['name'])) ? $_FILES['imagen']['name'] : "";
-        $fecha_imagen = new DateTime();
-        $nombre_archivo_imagen = ($imagen != "") ? $fecha_imagen->getTimestamp() . "_" . $imagen : "";
-
-        $tmp_imagen = $_FILES["imagen"]["tmp_name"];
-
-        move_uploaded_file($tmp_imagen, "../../../assets/img/team/" . $nombre_archivo_imagen);
-        
-        //borrado del archivo anterior
-        $sentencia = $conexion->prepare("SELECT imagen FROM `tbl_equipo` WHERE ID = :ID");
+        $sentencia->bindParam(':nombrecompleto', $nombrecompleto);
+        $sentencia->bindParam(':puesto', $puesto);
+        $sentencia->bindParam(':twitter', $twitter);
+        $sentencia->bindParam(':facebook', $facebook);
+        $sentencia->bindParam(':linkedin', $linkedin);
         $sentencia->bindParam(':ID', $txtID);
         $sentencia->execute();
-        $registro = $sentencia->fetch(PDO::FETCH_LAZY);
+        $mensaje = "Datos actualizados correctamente";
 
-        if (isset($registro["imagen"])) {
-            if(file_exists("../../../assets/img/team/" . $registro["imagen"])) {
-                unlink("../../../assets/img/team/" . $registro["imagen"]);
+        if ($_FILES['imagen']['tmp_name'] != "") {
+            $fecha_imagen = new DateTime();
+            $nombre_archivo_imagen = ($imagen != "") ? $fecha_imagen->getTimestamp() . "_" . $imagen : "";
+
+            $tmp_imagen = $_FILES["imagen"]["tmp_name"];
+
+            move_uploaded_file($tmp_imagen, "../../../assets/img/team/" . $nombre_archivo_imagen);
+            
+            // Borrado del archivo anterior
+            $sentencia = $conexion->prepare("SELECT imagen FROM `tbl_equipo` WHERE ID = :ID");
+            $sentencia->bindParam(':ID', $txtID);
+            $sentencia->execute();
+            $registro = $sentencia->fetch(PDO::FETCH_LAZY);
+
+            if (isset($registro["imagen"])) {
+                if(file_exists("../../../assets/img/team/" . $registro["imagen"])) {
+                    unlink("../../../assets/img/team/" . $registro["imagen"]);
+                }
             }
+
+            $sentencia = $conexion->prepare("UPDATE `tbl_equipo` SET `imagen` = :imagen WHERE `ID` = :ID;");
+            $sentencia->bindParam(':imagen', $nombre_archivo_imagen);
+            $sentencia->bindParam(':ID', $txtID);
+            $sentencia->execute();
+            $imagen = $nombre_archivo_imagen;
+            $mensaje = "Datos actualizados correctamente";
         }
 
-        $sentencia = $conexion->prepare("UPDATE `tbl_equipo` SET `imagen` = :imagen WHERE `ID` = :ID;");
-        $sentencia->bindParam(':imagen', $nombre_archivo_imagen);
-        $sentencia->bindParam(':ID', $txtID);
-        $sentencia->execute();
-        $imagen=$nombre_archivo_imagen;
-        $mensaje = "Datos agregados correctamente";
         header("Location: index.php?mensaje=$mensaje");
     }
-    
 }
 
-include("../../templates/headers.php"); ?>
-
+include("../../templates/headers.php");
+?>
 
 <div class="card">
     <div class="card-header">
         Datos del Equipo
     </div>
     <div class="card-body">
-
         <form action="" method="post" enctype="multipart/form-data">
-
-        <div class="mb-3">
-          <label for="" class="form-label">ID:</label>
-          <input type="text"
-            class="form-control" readonly value="<?php echo $txtID;?>" name="txtID" id="txtID" aria-describedby="helpId" placeholder="ID">
-         </div>
+            <div class="mb-3">
+                <label for="txtID" class="form-label">ID:</label>
+                <input type="text" class="form-control" readonly value="<?php echo $txtID; ?>" name="txtID" id="txtID" aria-describedby="helpId" placeholder="ID">
+            </div>
 
             <div class="mb-3">
-                <label for="imagen" class="form-label">Imagen: </label>
+                <label for="imagen" class="form-label">Imagen:</label>
                 <img width="50" src="../../../assets/img/team/<?php echo $imagen; ?>" />
                 <input type="file" class="form-control" name="imagen" id="imagen" aria-describedby="helpId" placeholder="Imagen">
             </div>
@@ -113,19 +111,23 @@ include("../../templates/headers.php"); ?>
 
             <div class="mb-3">
                 <label for="facebook" class="form-label">Facebook:</label>
-                <input type="text" class="form-control" value="<?php echo $facebook; ?>" name="facebook" id="facebook" aria-describedby="helpId" placeholder="Faceboook">
+                <input type="text" class="form-control" value="<?php echo $facebook; ?>" name="facebook" id="facebook" aria-describedby="helpId" placeholder="Facebook">
             </div>
 
             <div class="mb-3">
-                <label for="linkedin" class="form-label">linkedin:</label>
-                <input type="text" class="form-control" value="<?php echo $linkedin; ?>" name="linkedin" id="linkedin" aria-describedby="helpId" placeholder="linkedin">
+                <label for="linkedin" class="form-label">LinkedIn:</label>
+                <input type="text" class="form-control" value="<?php echo $linkedin; ?>" name="linkedin" id="linkedin" aria-describedby="helpId" placeholder="LinkedIn">
             </div>
+
+            <?php if (isset($mensaje)) { ?>
+                <div class="alert alert-danger" role="alert">
+                    <?php echo $mensaje; ?>
+                </div>
+            <?php } ?>
 
             <button type="submit" class="btn btn-success">Actualizar</button>
             <a name="" id="" class="btn btn-primary" href="index.php" role="button">Cancelar</a>
-
         </form>
-
     </div>
     <div class="card-footer text-muted">
         Footer
